[tower] Allow tower to mine dynamic VDF parameters (#1111)
* add vdf difficulty type in rust

* refactor fetching local block

* refactor mine once

* refactor mine_once

* remove depr tests

* tests passing

* move next_proof to own module

* move vdf difficulty type

* function to create next proof params from chain query

* switch between local and online mode for proof generation

* move code for allowing vdf praams to be stored instead  of hard coded.

* add toy_rng for tower state

* update VDF params at end of epoch

* check correct security in period

* cleanup

* patch tower cli build

* patch toy_rng for tests

* patch tower app test

* build stdlib

* test passing

* WIP make new miner methods work with swarm

* tower can get difficulty params from chain

* genesis to use test settings

* fix loop and backlog issue

* don't init difficulty at genesis

* reorder genesis tx so testnet is initialized before genesis settings

* patch merge

* patch test

* new towers will start from the baseline/genesis difficulty and security params

* refactor default vdf params for block zero

* update tests for toy_rng

* test epoch creates random difficulty

* catch error with len == 0

* bad makefile

* add test for continuous mining

* scaffold gc for tower

* scaffold

* find first discontinous proof

* collect subsequent proofs function

* delete files

* clean

* backlog checks a bad tx for garbage collection

* implement GC on failed proof

* while transitioning to dynamic vdf, tower should failover to localmode

* downgrade error msg to info

* cleanup unused prometheus log code

* downgrade state sync error message to info

* makefile for set-waypoint

* patch makefile

* refresh onchain state

* genesis should assign vouches

* set layout for devnet has 4 nodes

* init val in genesis first

* github repo tools can put file in genesis repo

* save files to genesis repo

* reorganize testnet scripts

* cleanup

* documentation

* documentation

* patch makefile

* patch makefile

* rename make recipes

* rename

* vdf difficulties must always be even numbers

* patch move

* makefile

* makefile

* text [skip-ci]

* update yaml configs so that node file uses multiple fullnode networks [skip-ci]

* patch [skip ci]

* patch ports [skip ci]

* validator init on genesis and ol init, will use seed peers

* trim whitespace on genesis_waypoint.txt

* makefile

* method to create  a genesis proof

* reverse state-sync default params

* add helper in init to reset safety data

* bump version

* ol restore has --boundary-only or -b which will not attempt to restore an advanced version. This is the safest restore option.

* ol restore defaults to only fetch the epoch boundary. --highest-version must be explicitly passed if the user wants versions beyond the epoch boundary (this is still experimental).

* ol serve: deprecating the --run-checks or -c option. We now assume the user wants to run checks while running the web-monitor

* improve error messages on safety rules initialize

* validator should panic on trying to startup safety rules

* patch

* makefile

* make config chain-id and instance of namedchain, breaking

* patch build

* make chainId::str_to_chain_id public

* parse namedchain from str

* use testing ID in tx_params

* make testnet start with alice, bob, carol only

* add execution.genesis_file_location to fullnode yaml

* set layout to have 5 addresses for testnet genesis

* alway use cargo release mode even in testnet

* undo

* update fixture files with ip addresses for reference.

* update check for prod settings for nextproof

* VDF native function was aborting instead of returning error.

* NextProof generate genesis proof from configs

* add derive traits to nextproof

* builds

* refactor get_next_proof_from_chain [skip-ci]

* patch test because chain_id options have changed

* patch testtepoch_random_difficulty.move

* refactor init and onboard to use new named chain property

* patch copy of fixture files in swarm

* patch wrong chainid for swarm

* patch test mining epochs

* patch chain_id build issues

* include more namedchain str parsing options

* tower: delete a vdf proof that cannot be parsed

* debug ci

+++ ol/tower/src/delay.rs
@@ -5,9 +5,9 @@ use anyhow::{Error, bail};
 use vdf::{PietrzakVDFParams, VDF, VDFParams};
 
 /// Runs the VDF
-pub fn do_delay(preimage: &[u8], difficulty: u64, security: u16) -> Result<Vec<u8>, Error> {
+pub fn do_delay(preimage: &[u8], difficulty: u64, security: u64) -> Result<Vec<u8>, Error> {
     // Functions for running the VDF.
-    let vdf: vdf::PietrzakVDF = PietrzakVDFParams(security).new();
+    let vdf: vdf::PietrzakVDF = PietrzakVDFParams(security as u16).new();
     match vdf.solve(preimage, difficulty) {
         Ok(proof) => Ok(proof),
         Err(e) => bail!(format!("ERROR: cannot solve VDF, message {:?}", e)),
